refactor(player): remove ngrx-effects for CreatePlayer() action
