+++ src/app/player/player.actions.ts
@@ -3,8 +3,6 @@ import { Action } from '@ngrx/store';
 export const UPDATE_PLAYER = '[player] UPDATE SUCCESS';
 export const UPDATE_PLAYER_SUCCESS = '[player] UPDATE PLAYER SUCCESS';
 export const UPDATE_PLAYER_FAIL = '[player] UPDATE PLAYER FAIL';
-export const CREATE_PLAYER = '[player] CREATE';
-export const CREATE_PLAYER_SUCCESS = '[player] CREATE SUCCESS';
 
 export class PlayerActions implements Action {
     constructor(public type: string = '', public payload: any = {}) {
@@ -21,12 +19,4 @@ export function UpdatePlayerSuccess(payload: any): PlayerActions {
 
 export function UpdatePlayerFail(): PlayerActions {
     return new PlayerActions(UPDATE_PLAYER_FAIL);
-}
-
-export function CreatePlayer(payload: any): PlayerActions {
-    return new PlayerActions(CREATE_PLAYER, payload);
-}
-
-export function CreatePlayerSuccess(payload: any): PlayerActions {
-    return new PlayerActions(CREATE_PLAYER_SUCCESS, payload);
 }
