+++ src/app/player/player.reducer.ts
@@ -1,6 +1,4 @@
 import {
-    CREATE_PLAYER,
-    CREATE_PLAYER_SUCCESS,
     PlayerActions,
     UPDATE_PLAYER,
     UPDATE_PLAYER_FAIL,
@@ -24,17 +22,6 @@ export function playerReducer(playerState: Player, action: PlayerActions) {
                 loading: false,
             };
         case UPDATE_PLAYER_FAIL:
-            return {
-                ...playerState,
-                ...action.payload,
-                loading: false,
-            };
-        case CREATE_PLAYER:
-            return {
-                ...playerState,
-                loading: true,
-            };
-        case CREATE_PLAYER_SUCCESS:
             return {
                 ...playerState,
                 ...action.payload,
